pyproject.toml
[tool.poetry]
name = "tfbs"
version = "0.1.0"
description = ""
authors = ["Malcolm Perry <malcolm@ebi.ac.uk>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
MOODS-python = "^1.9.4"
numpy = "^1.21.2"
click = "^8.1.3"
pyfaidx = "^0.7.1"
pybedtools = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipykernel = "^6.26.0"
polars = "^0.19.13"
ncls = "^0.0.68"
pandas = "^2.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
tfbs = "tfbs.cli:scan"
